# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ShiftManager API
  version: "1.0.0"
  description: REST API for the ShiftManager club volunteer-hours system.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

    # ── Member ────────────────────────────────────────────────────────────────
...
    AppSettings:
      type: object
      required: [clubName, yearGoal, clubYear, nameMode, reservationHours, billingMode, kioskSearch, feeSchedule, deregisterDeadlineH]
      properties:
        clubName:           { type: string }
        yearGoal:           { type: number }
        clubYear:           { type: string }
        nameMode:           { type: string, enum: [abbrev, full] }
        reservationHours:   { type: integer }
        billingMode:        { type: string, enum: [auto, manuell] }
        kioskSearch:        { type: boolean }
        feeSchedule:
          type: array
          items: { type: number }
        deregisterDeadlineH: { type: integer }
        kioskLocked:         { type: boolean }
        reminderLeadWeeks:   { type: integer }
    BrandingConfig:
      type: object
      required: [primaryColor, clubName]
      properties:
        primaryColor: { type: string }
        logoUrl:      { type: string, nullable: true }
        clubName:     { type: string }
    BrandingUpdateResult:
      type: object
      required: [branding]
      properties:
        branding:  { $ref: '#/components/schemas/BrandingConfig' }
        warnings:
          type: array
          items: { type: string }
...
    FeeTier:
      type: object
      required: [id, clubYearId, position, amountCents]
      properties:
        id:          { $ref: '#/components/schemas/UUID' }
        clubYearId:  { $ref: '#/components/schemas/UUID' }
        position:    { type: integer }
        amountCents: { type: integer }
    MemberFeeTier:
      type: object
      required: [id, clubYearId, position, amountCents]
      properties:
        id:          { $ref: '#/components/schemas/UUID' }
        memberId:    { $ref: '#/components/schemas/UUID' }
        clubYearId:  { $ref: '#/components/schemas/UUID' }
        position:    { type: integer }
        amountCents: { type: integer }

    # ── Audit ─────────────────────────────────────────────────────────────────
    AuditEntry:
      type: object
      required: [id, action, entity, entityId, createdAt]
      properties:
        id:        { $ref: '#/components/schemas/UUID' }
        actorId:   { $ref: '#/components/schemas/UUID', nullable: true }
        action:    { type: string }
        entity:    { type: string }
        entityId:  { type: string }
        before:    { type: object, nullable: true }
        after:     { type: object, nullable: true }
        createdAt: { type: string, format: date-time }

    # ── Email ─────────────────────────────────────────────────────────────────
    EmailTemplate:
      type: object
      required: [name, subject, body]
      properties:
        name:    { type: string }
        subject: { type: string }
        body:    { type: string }
    EmailLogEntry:
      type: object
      required: [id, to, subject, status, createdAt]
      properties:
        id:        { $ref: '#/components/schemas/UUID' }
        to:        { type: string, format: email }
        template:  { type: string }
        subject:   { type: string }
        body:      { type: string }
        status:    { type: string, enum: [sent, failed] }
        error:     { type: string }
        createdAt: { type: string, format: date-time }

    # ── Stats ─────────────────────────────────────────────────────────────────
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get application settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppSettings' }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/branding:
    get:
      operationId: getBranding
      tags: [settings]
      summary: Get club branding
      security: []
      responses:
        "200":
          description: Branding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrandingConfig' }
    put:
      operationId: updateBranding
      tags: [settings]
      summary: Update branding (Vorstand+); returns WCAG contrast warnings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrandingConfig' }
      responses:
        "200":
          description: Branding + warnings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrandingUpdateResult' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/fee-tiers:
    get:
      operationId: getFeeTiers
      tags: [settings]
      summary: Get fee tiers
      responses:
        "200":
          description: Fee tiers
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/FeeTier' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateFeeTiers
      tags: [settings]
      summary: Update fee tiers (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { $ref: '#/components/schemas/FeeTier' } }
      responses:
        "200":
          description: Updated tiers
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/FeeTier' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/audit:
    get:
      operationId: getAuditLog
      tags: [settings]
      summary: Get audit log (Vorstand+)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AuditEntry' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/logo:
    post:
      operationId: uploadLogo
      tags: [settings]
      summary: Upload club logo PNG/SVG (Vorstand+)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Updated branding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrandingConfig' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/email-templates:
    get:
      operationId: listEmailTemplates
      tags: [settings]
      summary: List email templates (Vorstand+)
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/EmailTemplate' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/email-templates/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    get:
      operationId: getEmailTemplate
      tags: [settings]
      summary: Get an email template (Vorstand+)
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailTemplate' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    put:
      operationId: updateEmailTemplate
      tags: [settings]
      summary: Update an email template (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailTemplate' }
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailTemplate' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/email-log:
    get:
      operationId: getEmailLog
      tags: [settings]
      summary: Get email send log (Vorstand+)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/EmailLogEntry' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/email-log/{id}/resend:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: resendEmail
      tags: [settings]
      summary: Resend a logged email (Vorstand+)
      responses:
        "200":
          description: Resent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /settings/members/{id}/fee-tiers:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      - { name: clubYearId, in: query, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getMemberFeeTiers
      tags: [settings]
      summary: Get per-member fee tier overrides (Vorstand+)
      responses:
        "200":
          description: Overrides
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MemberFeeTier' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    put:
      operationId: updateMemberFeeTiers
      tags: [settings]
      summary: Set per-member fee tier overrides (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tiers]
              properties:
                tiers: { type: array, items: { $ref: '#/components/schemas/MemberFeeTier' } }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MemberFeeTier' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Billing ──────────────────────────────────────────────────────────────────
...